refactor: improve getting user data
