+++ catcare_front/src/screens/OwnerOnboarding/OwnerOnboardingScreen.tsx
@@ -3,7 +3,7 @@ import { useForm } from 'react-hook-form'
 import { zodResolver } from '@hookform/resolvers/zod'
 import * as z from 'zod'
 import { useNavigate } from 'react-router-dom'
-import { useCallback, useContext } from 'react'
+import { useCallback, useContext, useEffect } from 'react'
 import { RouterPaths } from '@/router/RouterPathsMapper'
 import { Button } from '@/components/Button/Button'
 import { TextField } from '@mui/material'
@@ -30,27 +30,33 @@ function OwnerOnboardingScreen() {
   const { authState } = useContext(AuthContext)
 
   const handleFinish = useCallback(async (data: OwnerSchema) => {
-    // http://localhost:3000/api/profile/onboarding
     const body = {
       ...data,
       userId: authState.user.id
     }
 
-    const response = await axios.post(`${import.meta.env.VITE_CATCARE_SERVER_URL}/profile/onboarding`, body, {
-      headers: {
-        'Content-Type': 'application/json',
-        Authorization: `Bearer ${authState.token}`
-      }
-    })
+    // const response = await axios.post(`${import.meta.env.VITE_CATCARE_SERVER_URL}/profile/onboarding`, body, {
+    //   headers: {
+    //     'Content-Type': 'application/json',
+    //     Authorization: `Bearer ${authState.token}`
+    //   }
+    // })
 
-    if (response.status < 200 || response.status >= 300) {
-      toast.error('Não foi possível finalizar o cadastro.')
-      return
-    }
+    // if (response.status < 200 || response.status >= 300) {
+    //   toast.error('Não foi possível finalizar o cadastro.')
+    //   return
+    // }
 
     toast.success('Cadastro finalizado!')
     navigate(RouterPaths.HOME)
   }, [])
+
+  useEffect(() => {
+    if (authState.user?.onBoardingDone) {
+      navigate(RouterPaths.HOME)
+    }
+  }, [authState.user?.onBoardingDone])
+
   return (
     <OwnerOnboardingContainer>
       <Header>
