+++ catcare_front/src/screens/SitterOnboarding/SitterOnboardingScreen.styles.ts
@@ -1,6 +1,17 @@
 import { TextField } from '@mui/material'
 import styled from 'styled-components'
 
+const ITEM_HEIGHT = 48
+const ITEM_PADDING_TOP = 8
+export const MenuProps = {
+  PaperProps: {
+    style: {
+      maxHeight: ITEM_HEIGHT * 4.5 + ITEM_PADDING_TOP,
+      width: 250
+    }
+  }
+}
+
 export const SitterOnboardingContainer = styled.div`
   height: 100%;
   display: flex;
